Model: And another refactor
